requirements.txt
Flask==2.3.3
gunicorn==21.2.0
openai>=1.35.0
python-dotenv>=1.0.0

langgraph
streamlit>=1.35.0
